--- app/src/recordingStates.js
@@ -1,6 +1,0 @@
-// Recoding STATE
-export const RECORDING_INIT = null;
-export const RECORDING_START = 'start';
-export const RECORDING_STOP = 'stop';
-export const RECORDING_PAUSE = 'pause';
-export const RECORDING_RESUME = 'resume';
